fix(TS): Déclaration des paramètres au lieu de globales (#84)
Objectif: pouvoir renommer le paramètres de traitement `batchKey` en `fromBatchKey` (comme décidé avec @JazzyPierrot lors de notre pair coding) en minimisant le risque de regressions sur les autres APIs que `compact` utilisant également des paramètres de traitement venant de l'API.

Problème: pour l'instant, les paramètres de traitements sont définis comme des variables globales partagées avec les 3 traitements: `compact`, `public` et `reduce.algo2`. => Toute modification d'un paramètre impacterait ces 3 traitements.

Solution proposée: Au lieu de définir des variables globales partagées entre ces 3 traitements, déclaration de variables non globales dans les fichiers où elles sont nécessaires.

Prochaine étape: renommage du paramètre `batchKey` en `fromBatchKey`, pour le traitement `compact` seulement.
