+++ dbmongo/js/compact/complete_reporder.ts
@@ -1,5 +1,8 @@
 import "../globals.ts"
 
+// Paramètres globaux utilisés par "compact"
+declare const serie_periode: Date[]
+
 // complete_reporder ajoute une propriété "reporder" pour chaque couple
 // SIRET+période, afin d'assurer la reproductibilité de l'échantillonage.
 export function complete_reporder(
